+++ gostcrypto/gostcipher/gost_34_13_2015.py
@@ -8,9 +8,9 @@ Block cipher modes according to GOST 34.13-2015.
 
 The module implements the modes of operation of block encryption algorithms
 "magma" and "kuznechik", described in GOST 34.13-2015.  The module includes
-the base class GOST3413205, GOST3413205ecb, GOST3413205cbc, GOST3413205cfb,
-GOST3413205ofb and GOST3413205ctr.  In addition the module includes the
-GOSTCipherError class and several General functions.
+the base class GOST3413205, classes GOST3413205ecb, GOST3413205cbc,
+GOST3413205cfb, GOST3413205ofb and GOST3413205ctr.  In addition the module
+includes the GOSTCipherError class and several General functions.
 """
 
 from copy import deepcopy
@@ -788,6 +788,7 @@ class GOST34132015mac(GOST34132015):
     to the 'update()' method so far.
     - hexdigest(): calculating the Message authentication code of the data
     passed to the 'update()' method so far an return it of the hexadecimal.
+    - clear(): clearing the values of iterative cipher keys.
 
     Attributes
     - block_size: an integer value the internal block size of the cipher
@@ -917,7 +918,7 @@ class GOST34132015mac(GOST34132015):
 
         Exception
         - GOSTCipherError('GOSTCipherError: invalid message authentication code
-        size') in case of the invalid message authentication code size.
+        size'): in case of the invalid message authentication code size.
         """
         return self.digest(mac_size).hex()
 
@@ -938,11 +939,4 @@ class GOSTCipherError(Exception):
     - invalid message authentication code size.
     """
 
-    def __init__(self, msg: str) -> None:
-        """
-        Initialize exception.
-
-        Parameters
-        - msg: message to output when an exception occurs.
-        """
-        self.msg = msg
+    pass