+++ gostcrypto/gostpbkdf/r_50_1_111_2016.py
@@ -176,11 +176,4 @@ class GOSTPBKDFError(Exception):
     - invalid size of the derived key.
     """
 
-    def __init__(self, msg: str) -> None:
-        """
-        Initialize exception.
-
-        Parameters
-        - msg: message to output when an exception occurs.
-        """
-        self.msg = msg
+    pass